完善分页操作

+++ qn_system/src/main/webapp/WEB-INF/jsp/questionnaire_listAdmin.jsp
@@ -35,7 +35,11 @@ function ToPageNext() {
 	location.href = "/qn_system/ToPageIndex/"+index;
 }function deletequestionnaire(id){
 	if(confirm("您确认删除吗?")){
-		location.href="/qn_system/delete_questionnaire/"+id;
+		if(index != ""){
+			location.href="/qn_system/delete_questionnaire/"+id;
+		}else{
+			alert("请输入跳转页！");
+		}
 	}
 }
 </script>
